requirements.txt
numpy==1.24.2
openpyxl==3.1.2
Pillow==9.5.0